Making less Sheets API calls and increasing speed
Instead of making a read to the spreadsheet with each update to an individual cell, all values from certain sheets are read into a list of dictionaries. This lowers the chances of API exhaustion over the quota and increases the speed of the program. Creating the first pieces of the webpage(s).