Fix "Process Shift Data" errors
Fixes specifically how shifts that existed were marked as "skipped" because the bounds for a "real" clocked shift were too narrow. Also accounts for cases where a student forgets to sign out of a shift earlier in the day, then signs in later in the day so that new sign in is not counted as a sign out for the previous shift.