Streamline alerts under one HTML using Jinja
Alert type and alert message for bootstrap alerts in student_signin and staff_index are set in views.py. The function within Views then renders an HTML file, showing the alert with correct type and message, not loading the entire page, just the necessary elements, of student_signin and staff_index.