Remove prep_scan_shifts method
Adds on to PR #22

I realized this method never actually does anything. It gets called, but the method is not set to any variables and does not return anything, so nothing seems to be done within it that affects the shift analyzing. I tested both before and after removing it and things didn't seem to change at all.